Abstracts private methods out into SipowiczInternal class
